REFACTOR : ChatController, ChatResultResponse, ChatRoomController 분할
ChatController에 몰려있던 기능을 여러 컴포넌트로 분할
