+++ src/main/java/com/example/mzting/dto/ChatRoomRequest.java
@@ -1,4 +1,4 @@
-package com.example.mzting.DTO;
+package com.example.mzting.dto;
 
 
 import lombok.AllArgsConstructor;
