+++ src/main/java/com/example/mzting/service/CommentService.java
@@ -1,7 +1,7 @@
 package com.example.mzting.service;
 
-import com.example.mzting.DTO.CommentDTO;
-import com.example.mzting.DTO.LikeDislikeCountInfo;
+import com.example.mzting.dto.CommentDTO;
+import com.example.mzting.dto.LikeDislikeCountInfo;
 import com.example.mzting.entity.Comment;
 import com.example.mzting.entity.MBTIPosts;
 import com.example.mzting.repository.CommentRepository;
@@ -13,7 +13,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
